flask_microservice/requirements.txt
APScheduler==3.10.1
Flask==2.3.2
flask-socketio==5.2.0
mock==5.0.2
python_engineio==4.5.1
websockets==10.4
simple-websocket
flask-socketio==5.2.0
gevent
gevent-websocket
asgiref
gunicorn